+++ src/components/Category/CategoryList.js
@@ -1,32 +1,36 @@
 import React, { useState, useEffect } from 'react';
 import { GetAllCategory, DeleteCategory } from '../hooks/categoryHook';
 import ModalForm from './ModalForm';
+import Loading from "../share/Loading";
 
 const CategoriesList = () => {
   const [categories, setCategories] = useState([]);
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        // Make a GET request to the API endpoint
-        const response = await GetAllCategory();
-        // Update the state with the fetched data
-        setCategories(response);
-      } catch (error) {
-        console.error('Error fetching data:', error);
-      }
-    };
+  const [loading, setLoading] = useState(false);
+  const fetchData = async () => {
+    try {
+      let response = await GetAllCategory();
+      setCategories(response);
+    } catch (error) {
+      console.error('Error fetching data:', error);
+    }
+  };
 
+  useEffect(() => {
     fetchData();
   }, []);
 
   const handleDelete = async (id) => {
-    //UNDONE: Confirmation & FIX 404 ERROR
-    DeleteCategory(id);
-    window.location.reload();
+    if(window.confirm('Are you sure you want to delete this category?')){
+      setLoading(true);
+      await DeleteCategory(id);
+      fetchData();
+      setLoading(false);
+    }
   };
 
   return (
     <>
+      {loading && <Loading />}
       <div className=''>
         <ModalForm categories={categories} action="add" />
       </div>
@@ -53,7 +57,9 @@ const CategoriesList = () => {
 
                       <td className='col-md-2'>
                         <div class="d-flex gap-2">
-                        <ModalForm categories={categories} id={category._id} index={index} action="edit" />
+                        <ModalForm categories={categories} id={category._id} action="edit"
+                          fetchData={()=>fetchData()}
+                        />
                           <button
                             type="button"
                             class="btn btn-danger"
